+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import { Autocomplete } from "./components/Autocomplete"
+import { Autocomplete } from "./components"
 
 function App() {
