+++ frontend/src/components/css/Autocomplete.css
@@ -21,17 +21,23 @@
     width: 50%;
     margin: 0 10px;
     border: none;
-    font-size: 24px;
+    font-size: 23px;
     color: black;
 }
 #autocomplete-search > input:focus {
     outline: none;
 }
-
 #autocomplete-search > p:hover {
     cursor: pointer;
     opacity: 0.8;
     text-decoration: underline;
+}
+#autocomplete-search div:nth-child(1) {
+    text-align: center;
+}
+#autocomplete-search div:nth-child(1) span {
+    display: block;
+    font-size: 10px;
 }
 
 
@@ -85,15 +91,17 @@
 #autocomplete-results.not-empty {
     max-height: 450px;
     margin-top: 20px;
+}#autocomplete-results.empty {
+    max-height: 0px;
 }
 
 .autocomplete-result {
     width: 100%;
+    min-height: 90px;
     display: flex;
     align-items: center;
     justify-content: space-evenly;
     color: rgb(120,120,120);
-    padding: 5px;
 }
 .autocomplete-result.selected {
     background-color: rgba(200,200,200,0.4);
@@ -101,13 +109,32 @@
 .autocomplete-result:hover {
     cursor: pointer;
 }
+.autocomplete-result.active {
+    opacity: 0.6;
+    border-left: 5px solid rgba(100,100,100,0);;
+}
 .autocomplete-result > img {
     margin: 0 10px;
     height: 70px;
     width: 70px;
     border-radius: 5px;
 }
+.autocomplete-result > div {
+    flex: 1;
+}
 .autocomplete-result h4 {
     font-weight: 600;
+    padding: 0 5px;
     color: black;
+}
+.autocomplete-result p {
+    padding: 0 5px;
+    font-weight: 300;
+    font-size: 15px;
+}
+
+#autocomplete-no-results{
+    text-align: center;
+    padding: 5px;
+    font-size: 19px;
 }
