+++ frontend/src/components/index.ts
@@ -1,0 +1,3 @@
+export { Autocomplete } from './Autocomplete/Autocomplete';
+export { AutocompleteResult } from './Autocomplete/AutocompleteResult';
+export type { SuggestionResponse } from './Autocomplete/Autocomplete';
