+++ frontend/src/index.css
@@ -6,13 +6,12 @@
   margin: 0;
 }
 #app {
-  height: 100vh;
-  width: 100vw;
-  justify-content: center;
+  min-height: 100vh;
   background-color: rgb(224,223,225);
 }
 #ac-cont{
   display: flex;
   justify-content: center;
-  padding-top: 30px;
+  padding: 30px;
+  
 }
